Move context keys into a workbench contribution
